Physical operators

Spin chain operators

LindbladVectorizedTensors.exchange_interaction — Function

exchange_interaction(s1::Index, s2::Index; kwargs...)

Return an OpSum object encoding the Hamiltonian part $-i[H, –]$ of an exchange interaction between sites s1 and s2 term in a GKSL equation.

LindbladVectorizedTensors.exchange_interaction_adjoint — Function

exchange_interaction_adjoint(s1::Index, s2::Index; kwargs...)

Return an OpSum object encoding the adjoint of the Hamiltonian part $-i[H, –]$ of an exchange interaction between sites s1 and s2 term in a GKSL equation.

LindbladVectorizedTensors.exchange_interaction′ — Function

exchange_interaction_adjoint(s1::Index, s2::Index; kwargs...)

Return an OpSum object encoding the adjoint of the Hamiltonian part $-i[H, –]$ of an exchange interaction between sites s1 and s2 term in a GKSL equation.

LindbladVectorizedTensors.spin_chain — Function

spin_chain(
    freqs::Vector{<:Real},
    coups::Vector{<:Real},
    sites::Vector{<:Index}
)

Return an OpSum object encoding the Hamiltonian part $-i[H, –]$ of the GKSL equation for a spin chain of frequencies freqs and coupling constants coups, on sites.

LindbladVectorizedTensors.spin_chain_adjoint — Function

spin_chain_adjoint(
    freqs::Vector{<:Real},
    coups::Vector{<:Real},
    sites::Vector{<:Index}
)

Return an OpSum object encoding the adjoint of the Hamiltonian part $-i[H, –]$ of the GKSL equation for a spin chain of frequencies freqs and coupling constants coups, on sites.

LindbladVectorizedTensors.spin_chain′ — Function

spin_chain_adjoint(
    freqs::Vector{<:Real},
    coups::Vector{<:Real},
    sites::Vector{<:Index}
)

Return an OpSum object encoding the adjoint of the Hamiltonian part $-i[H, –]$ of the GKSL equation for a spin chain of frequencies freqs and coupling constants coups, on sites.

Pseudomode operators

LindbladVectorizedTensors.dissipator_loss — Function

dissipator_loss(n::Int)

Return an OpSum object representing a dissipator operator (as in a GKSL equation) on site n, associated to the jump operator A (the annihilation operator). The OpNames A and Adag must be defined for the site type in use.

LindbladVectorizedTensors.dissipator_gain — Function

dissipator_gain(n::Int)

Return an OpSum object representing a dissipator operator (as in a GKSL equation) on site n, associated to the jump operator Adag (the creation operator). The OpNames A and Adag must be defined for the site type in use.

LindbladVectorizedTensors.dissipator — Function

dissipator(n::Int, frequency::Real, temperature::Real)

Return an OpSum object which represents the dissipator in the GKSL equation for given frequency and temperature, i.e. with dissipation coefficient 1/(exp(freq/temp) - 1).

LindbladVectorizedTensors.mixedlindbladplus — Function

mixedlindbladplus(n1::Int, n2::Int)

Return on OpSum expression which represents a mixed dissipation operator, appearing in the equation for two pseudomodes, on sites at positions n1 and n2 in the system.

LindbladVectorizedTensors.mixedlindbladminus — Function

mixedlindbladminus(n1::Int, n2::Int)

Return on OpSum expression which represents a mixed dissipation operator, appearing in the equation for two pseudomodes, on sites at positions n1 and n2 in the system.