Physical operators

Spin chain operators

LindbladVectorizedTensors.exchange_interaction — Function

exchange_interaction(s1::Index, s2::Index; kwargs...)

Return an OpSum object encoding the Hamiltonian part $-i[H, –]$ of an exchange interaction between sites s1 and s2 term in a GKSL equation.

LindbladVectorizedTensors.exchange_interaction_adjoint — Function

exchange_interaction_adjoint(s1::Index, s2::Index; kwargs...)

Return an OpSum object encoding the adjoint of the Hamiltonian part $-i[H, –]$ of an exchange interaction between sites s1 and s2 term in a GKSL equation.

LindbladVectorizedTensors.exchange_interaction′ — Function

exchange_interaction_adjoint(s1::Index, s2::Index; kwargs...)

Return an OpSum object encoding the adjoint of the Hamiltonian part $-i[H, –]$ of an exchange interaction between sites s1 and s2 term in a GKSL equation.

LindbladVectorizedTensors.spin_chain — Function

spin_chain(
    freqs::Vector{<:Real},
    coups::Vector{<:Real},
    sites::Vector{<:Index}
)

Return an OpSum object encoding the Hamiltonian part $-i[H, –]$ of the GKSL equation for a spin chain of frequencies freqs and coupling constants coups, on sites.

LindbladVectorizedTensors.spin_chain_adjoint — Function

spin_chain_adjoint(
    freqs::Vector{<:Real},
    coups::Vector{<:Real},
    sites::Vector{<:Index}
)

Return an OpSum object encoding the adjoint of the Hamiltonian part $-i[H, –]$ of the GKSL equation for a spin chain of frequencies freqs and coupling constants coups, on sites.

LindbladVectorizedTensors.spin_chain′ — Function

spin_chain_adjoint(
    freqs::Vector{<:Real},
    coups::Vector{<:Real},
    sites::Vector{<:Index}
)

Return an OpSum object encoding the adjoint of the Hamiltonian part $-i[H, –]$ of the GKSL equation for a spin chain of frequencies freqs and coupling constants coups, on sites.

Pseudomode operators

LindbladVectorizedTensors.dissipator_loss — Function

dissipator_loss(n::Int)

Return an OpSum object representing a dissipator operator (as in a GKSL equation) on site n, associated to the jump operator A (the annihilation operator). The OpNames A and Adag must be defined for the site type in use.

LindbladVectorizedTensors.dissipator_gain — Function

dissipator_gain(n::Int)

Return an OpSum object representing a dissipator operator (as in a GKSL equation) on site n, associated to the jump operator Adag (the creation operator). The OpNames A and Adag must be defined for the site type in use.

LindbladVectorizedTensors.dissipator — Function

dissipator(n::Int, frequency::Real, temperature::Real)

Return an OpSum object which represents the dissipator in the GKSL equation for given frequency and temperature, i.e. with dissipation coefficient 1/(exp(freq/temp) - 1).

LindbladVectorizedTensors.mixedlindbladplus — Function

mixedlindbladplus(n1::Int, n2::Int)

Return on OpSum expression which represents a mixed dissipation operator, appearing in the equation for two pseudomodes, on sites at positions n1 and n2 in the system.

LindbladVectorizedTensors.mixedlindbladminus — Function

mixedlindbladminus(n1::Int, n2::Int)

Return on OpSum expression which represents a mixed dissipation operator, appearing in the equation for two pseudomodes, on sites at positions n1 and n2 in the system.

Markovian closure operators

LindbladVectorizedTensors.Closure — Type

A Closure object stores the parameters that define a Markovian closure, providing different ways to input them.

LindbladVectorizedTensors.closure — Function

closure(
    Ω::Real, K::Real,
    α::Matrix{<:Real}, β::Matrix{<:Real}, w::Matrix{<:Real}
)

Construct a Closure object with asymptotic frequency Ω and coupling K, with the universal parameter set given by α, β and w, each of which is assumed to be a two-column matrix with the real parts of the parameters in the first column and the imaginary parts in the second one.

closure(
    Ω::Real, K::Real,
    α::Vector{<:Complex}, β::Vector{<:Complex}, w::Vector{<:Complex}
)

Construct a Closure object with asymptotic frequency Ω and coupling K, with the universal parameter set given by α, β and w, which are assumed to be vectors of complex numbers.

Base.length — Method

length(mc::Closure)

Return the length of a Markovian closure, i.e. the number of pseudomodes.

LindbladVectorizedTensors.freqs — Method

freqs(mc::Closure)

Return the list of frequencies of the pseudomodes in the given Markovian closure.

LindbladVectorizedTensors.innercoups — Method

innercoups(mc::Closure)

Return the list of inner couplings of the pseudomodes in the Markovian closure mc, i.e. the coupling constant between the pseudomodes.

LindbladVectorizedTensors.outercoups — Method

outercoups(mc::Closure)

Return the list of outer couplings of the pseudomodes in the Markovian closure mc, i.e. the coupling constants with the last point of the TEDOPA chain.

LindbladVectorizedTensors.damps — Method

damps(mc::Closure)

Return the list of damping coefficients of the pseudomodes in the Markovian closure mc.

LindbladVectorizedTensors.freq — Method

freq(mc::Closure, j::Int)

Return the frequency of the j-th pseudomode of the Markovian closure mc.

LindbladVectorizedTensors.innercoup — Method

innercoup(mc::Closure, j::Int)

Return the coupling coefficient between the j-th and the j+1-th pseudomodes of the Markovian closure mc.

LindbladVectorizedTensors.outercoup — Method

outercoup(mc::Closure, j::Int)

Return the coupling coefficient of the j-th pseudomode in the Markovian closure mc, with the last point of the TEDOPA chain.

LindbladVectorizedTensors.damp — Method

damp(mc::Closure, j::Int)

Return the damping coefficient of the j-th pseudomode in the Markovian closure mc.

LindbladVectorizedTensors.closure_op — Method

closure_op(mc::Closure, sites::Vector{<:Index}, chain_edge_site::Int)

Return an OpSum object encoding the Markovian closure operators with parameters given by mc, on sites sites, linked to the main TEDOPA/thermofield chain on site chain_edge_site. This closure replaces a chain starting from an empty state.

LindbladVectorizedTensors.closure_op_adjoint — Method

closure_op_adjoint(
    mc::Closure, sites::Vector{<:Index}, chain_edge_site::Int, gradefactor::Int
)

Return an OpSum object encoding the adjoint of the Markovian closure operators with parameters given by mc, on sites sites, linked to the main TEDOPA/thermofield chain on site chain_edge_site. The integer gradefactor is the parity (1: even, -1: odd) of the operator subject to the time evolution. This closure replaces a chain starting from an empty state.

LindbladVectorizedTensors.closure_op′ — Function

closure_op_adjoint(
    mc::Closure, sites::Vector{<:Index}, chain_edge_site::Int, gradefactor::Int
)

Return an OpSum object encoding the adjoint of the Markovian closure operators with parameters given by mc, on sites sites, linked to the main TEDOPA/thermofield chain on site chain_edge_site. The integer gradefactor is the parity (1: even, -1: odd) of the operator subject to the time evolution. This closure replaces a chain starting from an empty state.

LindbladVectorizedTensors.filled_closure_op — Function

filled_closure_op(mc::Closure, sites::Vector{<:Index}, chain_edge_site::Int)

Return an OpSum object encoding the Markovian closure operators with parameters given by mc, on sites sites, linked to the main TEDOPA/thermofield chain on site chain_edge_site. This closure replaces a chain starting from a completely filled state.

LindbladVectorizedTensors.filled_closure_op_adjoint — Function

filled_closure_op_adjoint(
    mc::Closure, sites::Vector{<:Index}, chain_edge_site::Int, gradefactor::Int
)

Return an OpSum object encoding the Markovian closure operators with parameters given by mc, on sites sites, linked to the main TEDOPA/thermofield chain on site chain_edge_site. The integer gradefactor is the parity (1: even, -1: odd) of the operator subject to the time evolution. This closure replaces a chain starting from a completely filled state.

LindbladVectorizedTensors.filled_closure_op′ — Function

filled_closure_op_adjoint(
    mc::Closure, sites::Vector{<:Index}, chain_edge_site::Int, gradefactor::Int
)

Return an OpSum object encoding the Markovian closure operators with parameters given by mc, on sites sites, linked to the main TEDOPA/thermofield chain on site chain_edge_site. The integer gradefactor is the parity (1: even, -1: odd) of the operator subject to the time evolution. This closure replaces a chain starting from a completely filled state.